+++ crates/aiken-project/Cargo.toml
@@ -18,6 +18,7 @@ hex = "0.4.3"
 ignore = "0.4.18"
 itertools = "0.10.1"
 miette = { version = "5.3.0", features = ["fancy"] }
+owo-colors = "3.5.0"
 pallas = "0.16.0"
 pallas-traverse = "0.16.0"
 petgraph = "0.6.2"
