+++ crates/aiken/src/cmd/mod.rs
@@ -2,7 +2,6 @@ pub mod build;
 pub mod check;
 pub mod deps;
 pub mod docs;
-pub mod error;
 pub mod fmt;
 pub mod lsp;
 pub mod new;
